feat: change all service file names to lowercase #1574
