+++ web/src/pages/knowledge/model.ts
@@ -1,5 +1,5 @@
 import { IKnowledge } from '@/interfaces/database/knowledge';
-import kbService from '@/services/kbService';
+import kbService from '@/services/knowledge-service';
 import { DvaModel } from 'umi';
 
 export interface KnowledgeModelState {
